feat: bo integration all projects endpoint done, tests wip
