+++ packages/applications-service-api/__tests__/integration/applications.test.js
@@ -11,10 +11,12 @@ const { request } = require('../__data__/supertest');
 const { Op } = require('sequelize');
 
 const mockFindUnique = jest.fn();
+const mockProjectFindMany = jest.fn();
 jest.mock('../../src/lib/prisma', () => ({
 	prismaClient: {
 		project: {
-			findUnique: (query) => mockFindUnique(query)
+			findUnique: (query) => mockFindUnique(query),
+			findMany: (query) => mockProjectFindMany(query)
 		}
 	}
 }));
@@ -76,156 +78,165 @@ describe('/api/v1/applications', () => {
 	});
 
 	describe('get all applications', () => {
-		beforeEach(() => {
-			mockFindAndCountAll.mockResolvedValueOnce({
-				rows: APPLICATIONS_NI_FILTER_COLUMNS,
-				count: APPLICATIONS_NI_DB.length
+		describe('when backOfficeIntegration.applications.getAllApplications is false', () => {
+			config.backOfficeIntegration.applications.getAllApplications = false;
+			beforeEach(() => {
+				mockFindAndCountAll.mockResolvedValueOnce({
+					rows: APPLICATIONS_NI_FILTER_COLUMNS,
+					count: APPLICATIONS_NI_DB.length
+				});
+			});
+			it('happy path', async () => {
+				mockFindAndCountAll.mockResolvedValueOnce({
+					rows: APPLICATIONS_NI_DB,
+					count: APPLICATIONS_NI_DB.length
+				});
+
+				const response = await request.get('/api/v1/applications');
+
+				expect(response.status).toEqual(200);
+				expect(response.body).toEqual({
+					applications: APPLICATIONS_FO,
+					currentPage: 1,
+					itemsPerPage: 25,
+					totalItems: 5,
+					totalPages: 1,
+					totalItemsWithoutFilters: 5,
+					filters: APPLICATIONS_FO_FILTERS
+				});
+			});
+
+			it('with filters applied', async () => {
+				mockFindAndCountAll.mockResolvedValueOnce({
+					rows: [APPLICATIONS_NI_DB[2], APPLICATIONS_NI_DB[3], APPLICATIONS_NI_DB[4]],
+					count: 3
+				});
+
+				const filteredApplications = [APPLICATIONS_FO[2], APPLICATIONS_FO[3], APPLICATIONS_FO[4]];
+
+				const queryString = [
+					'stage=acceptance',
+					'stage=recommendation',
+					'region=eastern',
+					'region=north_west',
+					'sector=energy',
+					'sector=transport'
+				].join('&');
+
+				const response = await request.get(`/api/v1/applications?${queryString}`);
+
+				expect(mockFindAndCountAll).toBeCalledWith(
+					expect.objectContaining({
+						where: {
+							[Op.and]: [
+								{ Region: { [Op.in]: ['Eastern', 'North West'] } },
+								{ Stage: { [Op.in]: [2, 5] } },
+								{
+									[Op.or]: [{ Proposal: { [Op.like]: 'EN%' } }, { Proposal: { [Op.like]: 'TR%' } }]
+								}
+							]
+						}
+					})
+				);
+
+				expect(response.status).toEqual(200);
+				expect(response.body).toEqual({
+					applications: filteredApplications,
+					currentPage: 1,
+					itemsPerPage: 25,
+					totalItems: 3,
+					totalPages: 1,
+					totalItemsWithoutFilters: 5,
+					filters: APPLICATIONS_FO_FILTERS
+				});
+			});
+
+			it('with search term applied', async () => {
+				mockFindAndCountAll.mockResolvedValueOnce({
+					rows: [APPLICATIONS_NI_DB[0]],
+					count: 1
+				});
+				const filteredApplications = [APPLICATIONS_FO[0]];
+
+				const response = await request.get('/api/v1/applications?searchTerm=London%20Resort');
+
+				expect(mockFindAndCountAll).toBeCalledWith(
+					expect.objectContaining({
+						where: {
+							[Op.or]: [
+								{ ProjectName: { [Op.like]: '%London Resort%' } },
+								{ PromoterName: { [Op.like]: '%London Resort%' } }
+							]
+						}
+					})
+				);
+
+				expect(response.status).toEqual(200);
+				expect(response.body).toEqual({
+					applications: filteredApplications,
+					currentPage: 1,
+					itemsPerPage: 25,
+					totalItems: 1,
+					totalPages: 1,
+					totalItemsWithoutFilters: 5,
+					filters: APPLICATIONS_FO_FILTERS
+				});
+			});
+
+			it('with search term and filters applied', async () => {
+				mockFindAndCountAll.mockResolvedValueOnce({
+					rows: [APPLICATIONS_NI_DB[2], APPLICATIONS_NI_DB[3]],
+					count: 2
+				});
+				const filteredApplications = [APPLICATIONS_FO[2], APPLICATIONS_FO[3]];
+
+				const queryString = [
+					'stage=acceptance',
+					'stage=recommendation',
+					'region=eastern',
+					'region=north_west',
+					'sector=energy',
+					'sector=transport',
+					'searchTerm=Nuclear'
+				].join('&');
+
+				const response = await request.get(`/api/v1/applications?${queryString}`);
+
+				expect(mockFindAndCountAll).toBeCalledWith(
+					expect.objectContaining({
+						where: {
+							[Op.and]: [
+								{ Region: { [Op.in]: ['Eastern', 'North West'] } },
+								{ Stage: { [Op.in]: [2, 5] } },
+								{
+									[Op.or]: [{ Proposal: { [Op.like]: 'EN%' } }, { Proposal: { [Op.like]: 'TR%' } }]
+								}
+							],
+							[Op.or]: [
+								{ ProjectName: { [Op.like]: '%Nuclear%' } },
+								{ PromoterName: { [Op.like]: '%Nuclear%' } }
+							]
+						}
+					})
+				);
+
+				expect(response.status).toEqual(200);
+				expect(response.body).toEqual({
+					applications: filteredApplications,
+					currentPage: 1,
+					itemsPerPage: 25,
+					totalItems: 2,
+					totalItemsWithoutFilters: 5,
+					totalPages: 1,
+					filters: APPLICATIONS_FO_FILTERS
+				});
 			});
 		});
-		it('happy path', async () => {
-			mockFindAndCountAll.mockResolvedValueOnce({
-				rows: APPLICATIONS_NI_DB,
-				count: APPLICATIONS_NI_DB.length
+		describe('when backOfficeIntegration.applications.getAllApplications is true', () => {
+			beforeEach(() => {
+				config.backOfficeIntegration.applications.getAllApplications = true;
 			});
-
-			const response = await request.get('/api/v1/applications');
-
-			expect(response.status).toEqual(200);
-			expect(response.body).toEqual({
-				applications: APPLICATIONS_FO,
-				currentPage: 1,
-				itemsPerPage: 25,
-				totalItems: 5,
-				totalPages: 1,
-				totalItemsWithoutFilters: 5,
-				filters: APPLICATIONS_FO_FILTERS
-			});
-		});
-
-		it('with filters applied', async () => {
-			mockFindAndCountAll.mockResolvedValueOnce({
-				rows: [APPLICATIONS_NI_DB[2], APPLICATIONS_NI_DB[3], APPLICATIONS_NI_DB[4]],
-				count: 3
-			});
-
-			const filteredApplications = [APPLICATIONS_FO[2], APPLICATIONS_FO[3], APPLICATIONS_FO[4]];
-
-			const queryString = [
-				'stage=acceptance',
-				'stage=recommendation',
-				'region=eastern',
-				'region=north_west',
-				'sector=energy',
-				'sector=transport'
-			].join('&');
-
-			const response = await request.get(`/api/v1/applications?${queryString}`);
-
-			expect(mockFindAndCountAll).toBeCalledWith(
-				expect.objectContaining({
-					where: {
-						[Op.and]: [
-							{ Region: { [Op.in]: ['Eastern', 'North West'] } },
-							{ Stage: { [Op.in]: [2, 5] } },
-							{
-								[Op.or]: [{ Proposal: { [Op.like]: 'EN%' } }, { Proposal: { [Op.like]: 'TR%' } }]
-							}
-						]
-					}
-				})
-			);
-
-			expect(response.status).toEqual(200);
-			expect(response.body).toEqual({
-				applications: filteredApplications,
-				currentPage: 1,
-				itemsPerPage: 25,
-				totalItems: 3,
-				totalPages: 1,
-				totalItemsWithoutFilters: 5,
-				filters: APPLICATIONS_FO_FILTERS
-			});
-		});
-
-		it('with search term applied', async () => {
-			mockFindAndCountAll.mockResolvedValueOnce({
-				rows: [APPLICATIONS_NI_DB[0]],
-				count: 1
-			});
-			const filteredApplications = [APPLICATIONS_FO[0]];
-
-			const response = await request.get('/api/v1/applications?searchTerm=London%20Resort');
-
-			expect(mockFindAndCountAll).toBeCalledWith(
-				expect.objectContaining({
-					where: {
-						[Op.or]: [
-							{ ProjectName: { [Op.like]: '%London Resort%' } },
-							{ PromoterName: { [Op.like]: '%London Resort%' } }
-						]
-					}
-				})
-			);
-
-			expect(response.status).toEqual(200);
-			expect(response.body).toEqual({
-				applications: filteredApplications,
-				currentPage: 1,
-				itemsPerPage: 25,
-				totalItems: 1,
-				totalPages: 1,
-				totalItemsWithoutFilters: 5,
-				filters: APPLICATIONS_FO_FILTERS
-			});
-		});
-
-		it('with search term and filters applied', async () => {
-			mockFindAndCountAll.mockResolvedValueOnce({
-				rows: [APPLICATIONS_NI_DB[2], APPLICATIONS_NI_DB[3]],
-				count: 2
-			});
-			const filteredApplications = [APPLICATIONS_FO[2], APPLICATIONS_FO[3]];
-
-			const queryString = [
-				'stage=acceptance',
-				'stage=recommendation',
-				'region=eastern',
-				'region=north_west',
-				'sector=energy',
-				'sector=transport',
-				'searchTerm=Nuclear'
-			].join('&');
-
-			const response = await request.get(`/api/v1/applications?${queryString}`);
-
-			expect(mockFindAndCountAll).toBeCalledWith(
-				expect.objectContaining({
-					where: {
-						[Op.and]: [
-							{ Region: { [Op.in]: ['Eastern', 'North West'] } },
-							{ Stage: { [Op.in]: [2, 5] } },
-							{
-								[Op.or]: [{ Proposal: { [Op.like]: 'EN%' } }, { Proposal: { [Op.like]: 'TR%' } }]
-							}
-						],
-						[Op.or]: [
-							{ ProjectName: { [Op.like]: '%Nuclear%' } },
-							{ PromoterName: { [Op.like]: '%Nuclear%' } }
-						]
-					}
-				})
-			);
-
-			expect(response.status).toEqual(200);
-			expect(response.body).toEqual({
-				applications: filteredApplications,
-				currentPage: 1,
-				itemsPerPage: 25,
-				totalItems: 2,
-				totalItemsWithoutFilters: 5,
-				totalPages: 1,
-				filters: APPLICATIONS_FO_FILTERS
-			});
+			// TODO
 		});
 	});
 });
