--- packages/applications-service-api/__tests__/unit/repositories/project.repository.test.js
@@ -1,25 +1,0 @@
-const { getByCaseReference } = require('../../../src/repositories/project.repository');
-
-const mockFindUnique = jest.fn();
-jest.mock('../../../src/lib/prisma', () => ({
-	prismaClient: {
-		project: {
-			findUnique: (query) => mockFindUnique(query)
-		}
-	}
-}));
-
-describe('project repository', () => {
-	describe('getByCaseReference', () => {
-		const caseReference = 'EN010009';
-
-		it('calls findUnique with caseReference', async () => {
-			await getByCaseReference(caseReference);
-
-			expect(mockFindUnique).toBeCalledWith({
-				where: { caseReference: caseReference },
-				include: { applicant: true }
-			});
-		});
-	});
-});
